refactor: split insights endpoints into dedicated InsightsController
