+++ Pdmt.Api/Infrastructure/DateHelper.cs
@@ -1,0 +1,14 @@
+namespace Pdmt.Api.Infrastructure;
+
+internal static class DateHelper
+{
+    internal static DateTimeOffset GetMonday(DateTimeOffset date)
+    {
+        var daysToSubtract = ((int)date.DayOfWeek + 6) % 7;
+        var monday = date.Date.AddDays(-daysToSubtract);
+        return new DateTimeOffset(monday, TimeSpan.Zero);
+    }
+
+    internal static DateTimeOffset GetMonday(DateOnly date) =>
+        GetMonday(new DateTimeOffset(date.ToDateTime(TimeOnly.MinValue, DateTimeKind.Utc), TimeSpan.Zero));
+}
