+++ TODO.md
@@ -1,5 +1,5 @@
 # TODO — before Phase 4
 
 ## Рефакторинг
-- [ ] **`InsightsController` + `InsightsService`** — вынести 6 insights-методов из `AnalyticsController`/`AnalyticsService`; маршрут остаётся `api/analytics/insights/*`; `GetMonday` в статический хелпер
 - [ ] **MAUI: перевод на EN** — UI сейчас частично на русском
+- [ ] **Web: перевод на EN** — UI сейчас частично на русском
